FIX CI

+++ htdocs/core/lib/accounting.lib.php
@@ -383,7 +383,7 @@ function getDefaultDatesForTransfer()
 			}
 			$date_start = dol_mktime(0, 0, 0, $month_start, 1, $year_start);
 			$lastday = dol_get_last_day($year_end, $month_end);
-			$date_end = dol_mktime(23, 59, 59, $month_end, dol_print_date($lastday, '%d'), $year_end);
+			$date_end = dol_mktime(23, 59, 59, $month_end, (int) dol_print_date($lastday, '%d'), $year_end);
 		}
 	} elseif ($periodbydefaultontransfer == 1) {	// current month
 		$year_current = (int) dol_print_date(dol_now('gmt'), "%Y", 'gmt');
